Improve DockerCompose commands with Abstract class
