+++ app/Builder/DockerComposeExec.php
@@ -2,61 +2,17 @@
 
 namespace App\Builder;
 
-use App\Builder\Concerns\HasEnvironmentVariables;
-
-class DockerComposeExec extends Builder
+class DockerComposeExec extends DockerComposeAbstract
 {
-    use HasEnvironmentVariables;
-
-    /** @var string $user */
-    protected $user = '';
-
     public function getProgramName() : string
     {
         return 'exec';
     }
 
-    public function makeWrapper() : ?Builder
+    public function makeArgs(...$args) : array
     {
-        return new DockerCompose();
-    }
-
-    public function getDockerCompose() : DockerCompose
-    {
-        return $this->wrapper;
-    }
-
-    public function setUser(string $user) : self
-    {
-        $this->user = $user;
-
-        return $this;
-    }
-
-    public function getUser() : string
-    {
-        return $this->user;
-    }
-
-    protected function getEnvVar() : string
-    {
-        return env('FWD_COMPOSE_EXEC_FLAGS');
-    }
-
-    protected function beforeBuild(Builder $command) : Builder
-    {
-        $command->parseEnvironmentToArgument();
-
-        if ($user = $command->getUser()) {
-            $command->prependArgument(
-                new Argument('--user', Unescaped::make($user), ' ')
-            );
-        }
-
-        $command->prependArgument(new Argument(
-            Unescaped::make($this->getEnvVar())
-        ));
-
-        return $command;
+        return array_merge([
+            Unescaped::make(env('FWD_COMPOSE_EXEC_FLAGS')),
+        ], parent::makeArgs(...$args));
     }
 }
