+++ tests/TestCase.php
@@ -66,7 +66,6 @@ abstract class TestCase extends BaseTestCase
         $params = [
             env('FWD_DOCKER_COMPOSE_BIN', 'docker-compose'),
             sprintf('-p %s exec', basename(getcwd())),
-            env('FWD_COMPOSE_EXEC_FLAGS'),
         ];
 
         if (! empty($this->asUser)) {
@@ -74,6 +73,7 @@ abstract class TestCase extends BaseTestCase
             $params[] = $this->asUser;
         }
 
+        $params[] = env('FWD_COMPOSE_EXEC_FLAGS');
         $params[] = $this->buildCommand($command);
 
         $this->assertCommandRun($params);
@@ -138,23 +138,25 @@ abstract class TestCase extends BaseTestCase
         return "{$this->dockerComposeString()} exec";
     }
 
-    protected function makeDockerComposeRunString(string $args = ''): string
+    protected function makeDockerComposeRunString(string $args = '', string $envs = ''): string
     {
         $flags = env('FWD_COMPOSE_RUN_FLAGS') ? ' ' . env('FWD_COMPOSE_RUN_FLAGS') : '';
+        $extraArgs = $envs ? " -e $envs " : ' ';
 
-        return trim($this->dockerComposeRunString() . $flags . ' ' . $args);
+        return trim($this->dockerComposeRunString($extraArgs) . $flags . ' ' . $args);
     }
 
     protected function makeDockerComposeRunUserString($user = null, string $args = ''): string
     {
         $flags = env('FWD_COMPOSE_RUN_FLAGS') ? ' ' . env('FWD_COMPOSE_RUN_FLAGS') : '';
+        $extraArgs = " --user {$user} ";
 
-        return trim($this->dockerComposeRunString() . $flags . ' --user ' . $user . ' ' . $args);
+        return trim($this->dockerComposeRunString($extraArgs) . $flags . ' ' . $args);
     }
 
-    protected function dockerComposeRunString(): string
+    protected function dockerComposeRunString(string $extraArgs = ' '): string
     {
-        return "{$this->dockerComposeString()} run --rm";
+        return "{$this->dockerComposeString()} run" . $extraArgs . "--rm";
     }
 
     protected function dockerComposeString(): string
