+++ tests/fixtures/.env.docker-compose_run
@@ -1,2 +1,2 @@
 # Docker Compose flags
-FWD_COMPOSE_RUN_FLAGS="--rm -it"
+FWD_COMPOSE_RUN_FLAGS="-T"
